# openapi.yaml
openapi: 3.0.0
info:
  title: Taska Backend API
  description: REST API для платформы Taska
  version: 1.0.0
  contact:
    name: Taska Support
    email: support@taska.com
servers:
  - url: http://localhost:3000
    description: Local server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        message:
          type: string
          description: Error message
        errors:
          type: array
          items:
            type: object
            properties:
              msg:
                type: string
              param:
                type: string
              location:
                type: string
...
    Document:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        type:
          type: string
          enum: [contract, specification, report, other]
        fileType:
          type: string
        size:
          type: integer
        path:
          type: string
        uploadedBy:
          type: string
        uploadedAt:
          type: string
          format: date-time
        uploader:
          type: object
          properties:
            id:
              type: string
            fullName:
              type: string
            profileImage:
              type: string
              nullable: true
...
paths:
...
  /api/projects/{projectId}/documents:
    get:
      summary: Получить список документов проекта
      tags:
        - Documents
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: projectId
          schema:
            type: string
          description: ID проекта
        - in: query
          name: type
          schema:
            type: string
            enum: [contract, specification, report, other]
          description: Фильтр по типу документа
        - in: query
          name: uploadedBy
          schema:
            type: string
          description: Фильтр по загрузившему
        - in: query
          name: limit
          schema:
            type: integer
          description: Лимит документов
        - in: query
          name: before
          schema:
            type: string
          description: Курсор для пагинации
      responses:
        '200':
          description: Список документов
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Document'
        '401':
          description: Неавторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Нет доступа к проекту
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Проект не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      summary: Загрузить документ
      tags:
        - Documents
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: projectId
          schema:
            type: string
          description: ID проекта
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
                - name
              properties:
                file:
                  type: string
                  format: binary
                  description: Файл для загрузки (PDF, DOC, DOCX, XLS, XLSX, JPEG, PNG)
                name:
                  type: string
                  description: Название документа
                description:
                  type: string
                  description: Описание документа
                type:
                  type: string
                  enum: [contract, specification, report, other]
                  description: Тип документа
      responses:
        '201':
          description: Документ загружен
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  documentId:
                    type: string
        '400':
          description: Ошибка валидации или неверный тип файла
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Неавторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Нет доступа к проекту
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Проект не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...